fix: extract some board functions
